modularized the app, built a simple server that hosts on PORT 3000, next to deploy to Heroku.
