+++ src/css/style.css
@@ -1,0 +1,3 @@
+form {
+  background-color: aqua;
+}
