+++ webpack.config.js
@@ -11,6 +11,10 @@ module.exports = {
         },
       },
       {
+        test: /\.css$/i,
+        use: ["css-loader"],
+      },
+      {
         test: /\.html$/,
         use: [
           {
